+++ mq.sh
@@ -21,12 +21,33 @@ stop()
   endmqm $MQ_QMGR_NAME
 }
 
-config()
+parameterCheck()
 {
   : ${MQ_QMGR_NAME?"ERROR: You need to set the MQ_QMGR_NAME environment variable"}
+
+  # We want to do parameter checking early as then we can stop and error early before it looks
+  # like everything is going to be ok (when it won't)
+  if [ ! -z ${MQ_TLS_KEYSTORE+x} ]; then
+    if [ -z ${MQ_TLS_PASSPHRASE+x} ]; then
+      echo "Error: If you supply MQ_TLS_KEYSTORE, you must supply MQ_TLS_PASSPHRASE"
+      exit 1;
+    fi
+  fi
+}
+
+config()
+{
   # Populate and update the contents of /var/mqm - this is needed for
 	# bind-mounted volumes, and also to migrate data from previous versions of MQ
-  /opt/mqm/bin/amqicdir -i -f
+
+  setup-var-mqm.sh
+
+  if [ -z "${MQ_DISABLE_WEB_CONSOLE}" ]; then
+    echo $MQ_ADMIN_PASSWORD
+    # Start the web console, if it's been installed
+    which strmqweb && setup-mqm-web.sh
+  fi
+
   ls -l /var/mqm
   source /opt/mqm/bin/setmqenv -s
   echo "----------------------------------------"
@@ -38,6 +59,11 @@ config()
     echo "Checking filesystem..."
     amqmfsck /var/mqm
     echo "----------------------------------------"
+    MQ_DEV=${MQ_DEV:-"true"}
+    if [ "${MQ_DEV}" == "true" ]; then
+      # Turns on early adopt if we're using Developer defaults
+      export AMQ_EXTRA_QM_STANZAS=Channels:ChlauthEarlyAdopt=Y
+    fi
     crtmqm -q ${MQ_QMGR_NAME} || true
     if [ ${MQ_QMGR_CMDLEVEL+x} ]; then
       # Enables the specified command level, then stops the queue manager
@@ -46,11 +72,16 @@ config()
     echo "----------------------------------------"
   fi
   strmqm ${MQ_QMGR_NAME}
+
+  # Turn off script failing here because of listeners failing the script
+  set +e
   for MQSC_FILE in $(ls -v /etc/mqm/*.mqsc); do
     runmqsc ${MQ_QMGR_NAME} < ${MQSC_FILE}
   done
-  # Start the web console, if it's been installed
-  which strmqweb && su mqm -c "bash strmqweb &"
+  set -e
+
+  echo "----------------------------------------"
+  mq-dev-config.sh ${MQ_QMGR_NAME}
   echo "----------------------------------------"
 }
 
@@ -66,6 +97,8 @@ monitor()
     sleep 1
   done
   dspmq
+
+  echo "IBM MQ Queue Manager ${MQ_QMGR_NAME} is now fully running"
 
   # Loop until "dspmq" says the queue manager is not running any more
   until [ "`state`" != "RUNNING" ]; do
@@ -85,6 +118,7 @@ monitor()
 }
 
 mq-license-check.sh
+parameterCheck
 config
 trap stop SIGTERM SIGINT
 monitor