+++ test/test.js
@@ -62,11 +62,12 @@ describe('MQ Docker sample', function() {
       exec(cmd, function (err, stdout, stderr) {
         if (err) reject(err);
         let containerId = stdout.trim();
+        let startStr = `IBM MQ Queue Manager ${QMGR_NAME} is now fully running`;
         // Run dspmq every second, until the queue manager comes up
         let timer = setInterval(function() {
-          exec(`docker exec ${containerId} dspmq -n`, function (err, stdout, stderr) {
+          exec(`docker logs ${containerId}`, function (err, stdout, stderr) {
             if (err) reject(err);
-            if (stdout && stdout.includes("RUNNING")) {
+            if (stdout && stdout.includes(startStr)) {
               // Queue manager is up, so clear the timer
               clearInterval(timer);
               exec(`docker inspect --format '{{ .NetworkSettings.Networks.${DOCKER_NETWORK}.IPAddress }}' ${containerId}`, function (err, stdout, stderr) {
@@ -142,13 +143,13 @@ describe('MQ Docker sample', function() {
     });
 
     describe('and an empty bind-mounted volume', function() {
+      this.timeout(20000);
       let volumeDir = null;
       before(function() {
         volumeDir = fs.mkdtempSync(VOLUME_PREFIX);
       });
 
       before(function() {
-        this.timeout(20000);
         return runContainer(`--volume ${volumeDir}:/var/mqm`)
         .then((details) => {
           container = details;
@@ -176,6 +177,7 @@ describe('MQ Docker sample', function() {
     // This can happen if the entire volume directory is actually a filesystem.
     // See https://github.com/ibm-messaging/mq-docker/issues/29
     describe('and a non-empty bind-mounted volume', function() {
+      this.timeout(20000);
       let volumeDir = null;
       before(function() {
         volumeDir = fs.mkdtempSync(VOLUME_PREFIX);
@@ -183,7 +185,6 @@ describe('MQ Docker sample', function() {
       });
 
       before(function() {
-        this.timeout(20000);
         return runContainer(`--volume ${volumeDir}:/var/mqm`)
         .then((details) => {
           container = details;
@@ -204,6 +205,79 @@ describe('MQ Docker sample', function() {
         const client = net.connect({host: container.addr, port: 1414}, () => {
             client.on('close', done);
             client.end();
+        });
+      });
+    });
+
+    // Bluemix presented a new challenge when attempting to use it's volumes
+    // Because Bluemix is more secure and only allows root users to edit mounted
+    // volumes (and MQ runs as mqm) we have to mount the volume at another directory
+    // and then mount /var/mqm to a folder within that which has the correct owner.
+    describe('and a mounted volume in a different location', function() {
+      this.timeout(20000);
+      let volumeDir = null;
+      before(function() {
+        volumeDir = fs.mkdtempSync(VOLUME_PREFIX);
+      });
+
+      before(function() {
+        return runContainer(`--volume ${volumeDir}:/mnt/mqm`)
+        .then((details) => {
+          container = details;
+        });
+      });
+
+      after(function(done) {
+        deleteContainer(container.id)
+        .then(() => {
+          exec(`rm -rf  ${volumeDir}`, function (err, stdout, stderr) {
+            if (err) throw err;
+            done();
+          });
+        });
+      });
+
+      it('should be listening on port 1414 on the Docker network', function (done) {
+        const client = net.connect({host: container.addr, port: 1414}, () => {
+            client.on('close', done);
+            client.end();
+        });
+      });
+    });
+
+    // Test that we can stop and start a container without it failing. This Test
+    // makes sure that the scripts we set to run on every start can be ran when
+    // MQ data is already present.
+    describe('and can be started multiple times', function() {
+      this.timeout(20000);
+      before(function() {
+        return runContainer(``)
+        .then((details) => {
+          container = details;
+        });
+      });
+
+      after(function() {
+        return deleteContainer(container.id);
+      });
+
+      it('should not fail', function (done) {
+        exec(`docker stop ${container.id}`, function (err, stdout, stderr) {
+          if (err) throw err;
+          exec(`docker start ${container.id}`, function (err, stdout, stderr) {
+            if (err) throw err;
+            let startStr = `IBM MQ Queue Manager ${QMGR_NAME} is now fully running`;
+            let timer = setInterval(function() {
+              exec(`docker logs ${container.id}`, function (err, stdout, stderr) {
+                if (err) throw(err);
+                if (stdout && stdout.includes(startStr)) {
+                  // Queue manager is up, so clear the timer
+                  clearInterval(timer);
+                  done();
+                }
+              });
+            }, 1000);
+          });
         });
       });
     });